--- config/adc_rules.cfg
@@ -1,51 +1,0 @@
-; SPDX-FileCopyrightText: 2020-present Open Networking Foundation <info@opennetworking.org>
-;
-; SPDX-License-Identifier: Apache-2.0
-; SPDX-License-Identifier: LicenseRef-ONF-Member-Only-1.0
-
-[GLOBAL]
-NUM_ADC_RULES = 5
-
-;FORMAT ::
-;ADC_TYPE : [ DOMAIN = 0 | IP = 1 | IP PREFIX =2 ]
-;
-;if ADC_TYPE = 0
-;   DOMAIN
-;elseif ADC_TYPE = 1
-;   IP
-;elseif ADC_TYPE = 2
-;   IP
-;   PREFIX
-;else
-;   NONE
-;
-;NOTE :
-;Rules defined first have a higher priority, unless DROP is specified
-;(i.e. multiple rules for the same IP).
-;When specifying DROP with an IP address, use a prefix of 32 to prevent DNS
-;results from overwriting rule.
-
-
-[ADC_RULE_1]
-ADC_TYPE = 1
-IP = 13.2.1.111
-
-[ADC_RULE_2]
-ADC_TYPE = 2
-IP = 13.2.1.0
-PREFIX = 24
-
-[ADC_RULE_3]
-ADC_TYPE = 1
-IP = 13.1.1.112
-PREFIX = 24
-
-[ADC_RULE_4]
-ADC_TYPE = 0
-DOMAIN = www.example.gov
-
-[ADC_RULE_5]
-ADC_TYPE = 0
-DOMAIN = www.drop_example.com
-
-
