--- config/sdf_rules.cfg
@@ -1,44 +1,0 @@
-; SPDX-FileCopyrightText: 2020-present Open Networking Foundation <info@opennetworking.org>
-;
-; SPDX-License-Identifier: Apache-2.0
-; SPDX-License-Identifier: LicenseRef-ONF-Member-Only-1.0
-
-[GLOBAL]
-NUM_SDF_FILTERS = 4
-
-[SDF_FILTER_1]
-DIRECTION = downlink_only
-IPV4_REMOTE = 13.2.1.113
-IPV4_REMOTE_MASK = 255.255.255.0
-PROTOCOL = 17
-LOCAL_LOW_LIMIT_PORT = 0
-LOCAL_HIGH_LIMIT_PORT = 65535
-REMOTE_LOW_LIMIT_PORT = 0
-REMOTE_HIGH_LIMIT_PORT = 65535
-
-[SDF_FILTER_2]
-DIRECTION = uplink_only
-IPV4_LOCAL = 16.255.255.0
-IPV4_LOCAL_MASK = 255.255.255.0
-PROTOCOL = 17
-LOCAL_LOW_LIMIT_PORT = 0
-LOCAL_HIGH_LIMIT_PORT = 65535
-REMOTE_LOW_LIMIT_PORT = 0
-REMOTE_HIGH_LIMIT_PORT = 65535
-
-[SDF_FILTER_3]
-DIRECTION = downlink_only
-IPV4_REMOTE = 130.10.0.0
-IPV4_REMOTE_MASK = 255.255.0.0
-PROTOCOL = 17
-REMOTE_LOW_LIMIT_PORT = 5060
-REMOTE_HIGH_LIMIT_PORT = 5060
-
-[SDF_FILTER_4]
-DIRECTION = uplink_only
-IPV4_REMOTE = 103.1.0.0
-IPV4_REMOTE_MASK = 255.255.0.0
-PROTOCOL = 17
-LOCAL_LOW_LIMIT_PORT = 17000
-LOCAL_HIGH_LIMIT_PORT = 17010
-
