+++ cp/packet_filters.c
@@ -4,9 +4,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: LicenseRef-ONF-Member-Only-1.0
 
-#include <rte_acl.h>
-#include <rte_cfgfile.h>
-
 #include "ue.h"
 #include "util.h"
 #include "packet_filters.h"
@@ -60,670 +57,35 @@ uint64_t ebs;
 uint16_t ulambr_idx;
 uint16_t dlambr_idx;
 
-/**
- * @brief  : Converts char string to number
- * @param  : name , char string to be converted
- * @return : Returns converted number
- */
-static uint32_t name_to_num(const char *name)
-{
-	uint32_t num = 0;
-	int i;
-
-	for (i = strlen(name) - 1; i >= 0; i--)
-		num = (num << 4) | (name[i] - 'a');
-	return num;
-}
-
 int
 get_packet_filter_id(const pkt_fltr *pf)
 {
-	uint16_t index;
-	for (index = FIRST_FILTER_ID; index < num_packet_filters; ++index) {
-		if (!memcmp(pf, &packet_filters[index]->pkt_fltr,
-				sizeof(pkt_fltr)))
-			return index;
-	}
-	return -ENOENT;
+       uint16_t index;
+       for (index = FIRST_FILTER_ID; index < num_packet_filters; ++index) {
+               if (!memcmp(pf, &packet_filters[index]->pkt_fltr,
+                               sizeof(pkt_fltr)))
+                       return index;
+       }
+       return -ENOENT;
 }
 
 uint8_t
 get_packet_filter_direction(uint16_t index)
 {
-	return packet_filters[index]->pkt_fltr.direction;
+       return packet_filters[index]->pkt_fltr.direction;
 }
 
 packet_filter *
 get_packet_filter(uint16_t index)
 {
-	if (unlikely(index >= num_packet_filters))
-		return NULL;
-	return packet_filters[index];
+       if (unlikely(index >= num_packet_filters))
+               return NULL;
+       return packet_filters[index];
 }
 
 void
 reset_packet_filter(pkt_fltr *pf)
 {
-	memcpy(pf, &catch_all, sizeof(pkt_fltr));
-}
-
-int meter_profile_index_get(uint64_t cir)
-{
-	int index;
-	uint64_t CIR = cir >> 3; /* Convert bit rate into bytes */
-
-	for (index = 0; index < num_mtr_profiles; index++) {
-		if (mtr_profiles[index]->mtr_param.cir == CIR)
-			return mtr_profiles[index]->mtr_profile_index;
-	}
-
-	return 0;
-}
-
-void
-push_sdf_rules(uint16_t index)
-{
-	char local_ip[INET_ADDRSTRLEN];
-	char remote_ip[INET_ADDRSTRLEN];
-
-	snprintf(local_ip, sizeof(local_ip), "%s",
-	    inet_ntoa(sdf_filters[index]->local_ip_addr));
-	snprintf(remote_ip, sizeof(remote_ip), "%s",
-	    inet_ntoa(sdf_filters[index]->remote_ip_addr));
-
-	struct pkt_filter pktf = {
-			.pcc_rule_id = index
-	};
-
-	if (sdf_filters[index]->direction & TFT_DIRECTION_DOWNLINK_ONLY) {
-		snprintf(pktf.u.rule_str, MAX_LEN, "%s/%"PRIu8" %s/%"PRIu8
-			" %"PRIu16" : %"PRIu16" %"PRIu16" : %"PRIu16
-			" 0x%"PRIx8"/0x%"PRIx8"\n",
-			remote_ip, sdf_filters[index]->remote_ip_mask, local_ip,
-			sdf_filters[index]->local_ip_mask,
-			ntohs(sdf_filters[index]->remote_port_low),
-			ntohs(sdf_filters[index]->remote_port_high),
-			ntohs(sdf_filters[index]->local_port_low),
-			ntohs(sdf_filters[index]->local_port_high),
-			sdf_filters[index]->proto, sdf_filters[index]->proto_mask);
-		if (sdf_filters[index]->direction ==
-				TFT_DIRECTION_BIDIRECTIONAL)
-			LOG_MSG(LOG_ERROR, "Ignoring uplink portion of packet "
-					"filter for now");
-	} else if (sdf_filters[index]->direction & TFT_DIRECTION_UPLINK_ONLY) {
-		snprintf(pktf.u.rule_str, MAX_LEN, "%s/%"PRIu8" %s/%"PRIu8" %"
-			PRIu16" : %"PRIu16" %"PRIu16" : %"PRIu16" 0x%"
-			PRIx8"/0x%"PRIx8"\n",
-			local_ip, sdf_filters[index]->local_ip_mask, remote_ip,
-			sdf_filters[index]->remote_ip_mask,
-			ntohs(sdf_filters[index]->local_port_low),
-			ntohs(sdf_filters[index]->local_port_high),
-			ntohs(sdf_filters[index]->remote_port_low),
-			ntohs(sdf_filters[index]->remote_port_high),
-			sdf_filters[index]->proto, sdf_filters[index]->proto_mask);
-	}
-
-	LOG_MSG(LOG_DEBUG,"Installing %s pkt_filter #%"PRIu16" : %s",
-	    direction_str[sdf_filters[index]->direction], index,
-		pktf.u.rule_str);
-
-#ifdef OBSOLETE_APIS
-	struct dp_id dp_id = { .id = DPN_ID };
-	if (sdf_filter_entry_add(dp_id, pktf) < 0)
-        assert(0);
-#endif
-}
-
-/**
- * @brief  : Installs a sdf rules in the CP & DP.
- * @param  : new_packet_filter
- *           A sdf rules yet to be installed
- * @return : - >= 0 - on success - indicates index of sdf rules
- *           - < 0 - on error
- */
-static int
-install_sdf_rules(const pkt_fltr *new_packet_filter)
-{
-	if (num_sdf_filters >= SDF_FILTER_TABLE_SIZE)
-		return -ENOMEM;
-
-	pkt_fltr *filter = (pkt_fltr *)calloc(1, sizeof(pkt_fltr));
-	if (filter == NULL) {
-		LOG_MSG(LOG_ERROR, "Failure to allocate dedicated packet filter ");
-		return -ENOMEM;
-	}
-
-	memcpy(filter, new_packet_filter, sizeof(pkt_fltr));
-	uint16_t index = num_sdf_filters;
-
-	num_sdf_filters++;
-	sdf_filters[index] = filter;
-
-	push_sdf_rules(index);
-	return index;
-}
-
-/**
- * @brief  : Installs a pcc rules in the CP & DP.
- * @param  : new_pcc_entry
- *           A pcc rules yet to be installed
- * @return : - >= 0 - on success - indicates num_pcc_filter of pcc rules
- *           - < 0 - on error
- */
-static int
-install_pcc_rules(struct pcc_rules new_pcc_entry)
-{
-
-	if (num_pcc_filter >= PCC_TABLE_SIZE)
-		return -ENOMEM;
-
-	struct pcc_rules *pcc_filter = (struct pcc_rules *)calloc(1, sizeof(new_pcc_entry));
-	if (NULL == pcc_filter) {
-		LOG_MSG(LOG_ERROR, "Failure to allocate memeory for pcc filter ");
-		return -ENOMEM;
-	}
-
-	memcpy(pcc_filter, &new_pcc_entry, sizeof(new_pcc_entry));
-
-	pcc_filters[num_pcc_filter] = pcc_filter;
-	new_pcc_entry.rule_id = num_pcc_filter;
-	num_pcc_filter++;
-
-#ifdef OBSOLETE_APIS
-	struct dp_id dp_id = { .id = DPN_ID };
-	if (pcc_entry_add(dp_id, new_pcc_entry) < 0 )
-        assert(0);
-#endif
-	return num_pcc_filter;
-}
-
-/**
- * @brief  : Creates and adds meter profile entries
- * @param  : dp_id , dp id
- * @param  : new_mtr_entry, meter entry to be added
- * @return : Returns 0 in case of success , -1 otherwise
- */
-static int
-install_meter_profiles(struct dp_id dp_id, struct mtr_entry new_mtr_entry)
-{
-	if (num_mtr_profiles >= METER_PROFILE_SDF_TABLE_SIZE)
-		return -ENOMEM;
-
-	struct mtr_entry *mtr_profile = (struct mtr_entry *)calloc(1, sizeof(new_mtr_entry));
-	if (mtr_profile == NULL) {
-		LOG_MSG(LOG_ERROR, "Failure to allocate memeory for meter profile dp_id = %lu ", dp_id.id);
-		return -ENOMEM;
-	}
-
-	memcpy(mtr_profile, &new_mtr_entry, sizeof(new_mtr_entry));
-
-	mtr_profiles[num_mtr_profiles] = mtr_profile;
-	num_mtr_profiles++;
-
-#ifdef OBSOLETE_APIS
-	meter_profile_entry_add(dp_id, new_mtr_entry);
-#endif
-	return num_mtr_profiles;
-}
-
-/**
- * @brief  : Initialize meter profiles
- * @param  : No param
- * @return : Returns nothing
- */
-static void
-init_mtr_profile(void)
-{
-	unsigned no_of_idx = 0;
-	unsigned i = 0;
-	struct rte_cfgfile *file =
-			rte_cfgfile_load(METER_PROFILE_FILE, 0);
-	const char *entry;
-	struct dp_id dp_id = { .id = DPN_ID };
-
-    assert(file != NULL);
-
-	entry = rte_cfgfile_get_entry(file, "GLOBAL", "NUM_OF_IDX");
-    assert(entry != NULL);
-	no_of_idx = atoi(entry);
-
-	for (i = 1; i <= no_of_idx; ++i) {
-		char sectionname[64];
-		struct mtr_entry mtr_entry;
-
-		snprintf(sectionname, sizeof(sectionname),
-				"ENTRY_%u", i);
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"CIR");
-        assert(entry != NULL);
-		mtr_entry.mtr_param.cir = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"CBS");
-        assert(entry != NULL);
-		mtr_entry.mtr_param.cbs = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"EBS");
-        assert(entry != NULL);
-		mtr_entry.mtr_param.ebs = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"MTR_PROFILE_IDX");
-        assert(entry != NULL);
-		mtr_entry.mtr_profile_index = atoi(entry);
-
-		install_meter_profiles(dp_id, mtr_entry);
-
-	}
-}
-
-/**
- * @brief  : Initialize sdf rules
- * @param  : No param
- * @return : Returns nothing
- */
-static void
-init_sdf_rules(void)
-{
-	unsigned num_sdf_rules = 0;
-	unsigned i = 0;
-	const char *entry = NULL;
-	struct rte_cfgfile *file = rte_cfgfile_load(SDF_RULE_FILE, 0);
-
-    assert(file != NULL);
-
-	entry = rte_cfgfile_get_entry(file, "GLOBAL", "NUM_SDF_FILTERS");
-
-    assert(entry != NULL);
-
-	num_sdf_rules = atoi(entry);
-
-	for (i = 1; i <= num_sdf_rules; ++i) {
-		char sectionname[64] = {0};
-		int ret = 0;
-		struct in_addr tmp_addr;
-		pkt_fltr pf;
-
-		reset_packet_filter(&pf);
-		snprintf(sectionname, sizeof(sectionname),
-				"SDF_FILTER_%u", i);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "DIRECTION");
-		if (entry) {
-			if (strcmp(entry, "bidirectional") == 0){
-				pf.direction = TFT_DIRECTION_BIDIRECTIONAL;
-                assert(0); //Invalid SDF direction. Supported : uplink_only, downlink_only
-			}
-			else if (strcmp(entry, "uplink_only") == 0)
-				pf.direction = TFT_DIRECTION_UPLINK_ONLY;
-			else if (strcmp(entry, "downlink_only") == 0)
-				pf.direction = TFT_DIRECTION_DOWNLINK_ONLY;
-			else
-				assert(0); //Invalid SDF direction. Supported : uplink_only,downlink_only
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "IPV4_REMOTE");
-		if (entry) {
-			if (inet_aton(entry, &pf.remote_ip_addr) == 0)
-                assert(0);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"IPV4_REMOTE_MASK");
-		if (entry) {
-			ret = inet_aton(entry, &tmp_addr);
-			if (ret == 0
-			    || __builtin_clzl(~tmp_addr.s_addr)
-				+ __builtin_ctzl(tmp_addr.s_addr) != 32){
-				LOG_MSG(LOG_ERROR, "Invalid address %s in section %s "
-						"sdf config file %s. Setting to default 32.\n",
-						entry, sectionname, SDF_RULE_FILE);
-				 pf.remote_ip_mask = 32;
-			} else
-				pf.remote_ip_mask =
-						__builtin_popcountl(tmp_addr.s_addr);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"REMOTE_LOW_LIMIT_PORT");
-		if (entry)
-			pf.remote_port_low = htons((uint16_t) atoi(entry));
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"REMOTE_HIGH_LIMIT_PORT");
-		if (entry)
-			pf.remote_port_high = htons((uint16_t) atoi(entry));
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "PROTOCOL");
-		if (entry) {
-			pf.proto = atoi(entry);
-			pf.proto_mask = UINT8_MAX;
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "IPV4_LOCAL");
-		if (entry) {
-			if (inet_aton(entry, &pf.local_ip_addr) == 0)
-                assert(0);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"IPV4_LOCAL_MASK");
-		if (entry) {
-			ret = inet_aton(entry, &tmp_addr);
-			if (ret == 0
-			    || __builtin_clzl(~tmp_addr.s_addr)
-			    + __builtin_ctzl(tmp_addr.s_addr) != 32){
-				LOG_MSG(LOG_ERROR, "Invalid address %s in section %s "
-						"sdf config file %s. Setting to default 32.\n",
-						entry, sectionname, SDF_RULE_FILE);
-				pf.remote_ip_mask = 32;
-			} else
-				pf.local_ip_mask = __builtin_popcountl(tmp_addr.s_addr);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"LOCAL_LOW_LIMIT_PORT");
-		if (entry)
-			pf.local_port_low = htons((uint16_t) atoi(entry));
-
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"LOCAL_HIGH_LIMIT_PORT");
-		if (entry)
-			pf.local_port_high = htons((uint16_t) atoi(entry));
-
-		ret = install_sdf_rules(&pf);
-		if (ret < 0) {
-            assert(0);
-		}
-	}
-	num_sdf_filters = FIRST_FILTER_ID; /*Reset num_sdf_filters*/
-}
-
-/**
- * @brief  : Initialize pcc rules
- * @param  : No param
- * @return : Returns nothing
- */
-static void
-init_pcc_rules(void)
-{
-	unsigned num_pcc_rules = 0;
-	unsigned i = 0;
-	const char *entry = NULL;
-	struct rte_cfgfile *file = rte_cfgfile_load(PCC_RULE_FILE, 0);
-
-    assert(file != NULL);
-
-	entry = rte_cfgfile_get_entry(file, "GLOBAL", "NUM_PCC_FILTERS");
-
-    assert(entry != NULL);
-	num_pcc_rules = atoi(entry);
-
-	entry = rte_cfgfile_get_entry(file,
-				"GLOBAL", "UL_AMBR_MTR_PROFILE_IDX");
-    assert(entry != NULL);
-	ulambr_idx = atoi(entry);
-
-	entry = rte_cfgfile_get_entry(file,
-				"GLOBAL", "DL_AMBR_MTR_PROFILE_IDX");
-    assert(entry != NULL);
-	dlambr_idx = atoi(entry);
-
-	for (i = 1; i <= num_pcc_rules; ++i) {
-		char sectionname[64] = {0};
-		int ret = 0;
-		struct pcc_rules tmp_pcc = {0};
-
-		snprintf(sectionname, sizeof(sectionname),
-				"PCC_FILTER_%u", i);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "RULE_NAME");
-		if (entry)
-			strncpy(tmp_pcc.rule_name, entry, sizeof(tmp_pcc.rule_name));
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "RATING_GROUP");
-        assert(entry != NULL);
-		tmp_pcc.rating_group = atoi(entry);
-		if(0 == tmp_pcc.rating_group) {
-			tmp_pcc.rating_group = name_to_num(entry);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "SERVICE_ID");
-		if (entry) {
-			tmp_pcc.service_id = atoi(entry);
-			if(0 == tmp_pcc.service_id)
-				tmp_pcc.service_id = name_to_num(entry);
-		}
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "RULE_STATUS");
-		if (entry)
-			tmp_pcc.rule_status = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "GATE_STATUS");
-		if (entry)
-			tmp_pcc.gate_status = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "SESSION_CONT");
-		if (entry)
-			tmp_pcc.session_cont = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "REPORT_LEVEL");
-		if (entry)
-			tmp_pcc.report_level = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "CHARGING_MODE");
-		if (entry)
-			tmp_pcc.charging_mode = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "METERING_METHOD");
-		if (entry)
-			tmp_pcc.metering_method = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "MUTE_NOTIFY");
-		if (entry)
-			tmp_pcc.mute_notify = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "MONITORING_KEY");
-		if (entry)
-			tmp_pcc.monitoring_key = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "SPONSOR_ID");
-		if (entry)
-			strncpy(tmp_pcc.sponsor_id, entry, sizeof(tmp_pcc.sponsor_id));
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "REDIRECT_INFO");
-		if (entry)
-			tmp_pcc.redirect_info.info = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "PRECEDENCE");
-		if (entry)
-			tmp_pcc.precedence = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file, sectionname, "DROP_PKT_COUNT");
-		if (entry)
-			tmp_pcc.drop_pkt_count = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file,
-					sectionname, "UL_MBR_MTR_PROFILE_IDX");
-        assert(entry != NULL);
-
-		tmp_pcc.qos.ul_mtr_profile_index = atoi(entry);
-
-		entry = rte_cfgfile_get_entry(file,
-					sectionname, "DL_MBR_MTR_PROFILE_IDX");
-        assert(entry != NULL);
-		tmp_pcc.qos.dl_mtr_profile_index = atoi(entry);
-
-		/*Read mapped ADC or SDF rules. Either ADC or SDF rules will be p
-		 * resent, not both. SDF count will be 0 if ADC rules are present.*/
-		tmp_pcc.sdf_idx_cnt = 0;
-		entry = rte_cfgfile_get_entry(file,
-					sectionname, "ADC_FILTER_IDX");
-		if (!entry) {
-			/*No ADC entry, so check for SDF entry*/
-			entry = rte_cfgfile_get_entry(file,
-					sectionname, "SDF_FILTER_IDX");
-            assert(entry != NULL); // Missing SDF or ADC rule for PCC rule 
-
-			char *next = NULL;
-			uint16_t sdf_cnt = 0;
-			/*SDF entries format : "1, 2: 10, 30"*/
-			for(int x=0; x < MAX_SDF_IDX_COUNT; ++x) {
-				errno = 0;
-				int sdf_idx = strtol(entry, &next, 10);
-				if (errno != 0) {
-                    assert(0);
-				}
-				if('\0' == *entry) break;
-				/*If non number e.g.',', then ignore and continue*/
-				if(entry == next && (0 == sdf_idx)){
-					entry = ++next;
-					continue;
-				}
-				entry = next;
-				tmp_pcc.sdf_idx[sdf_cnt++] = sdf_idx;
-			}
-			tmp_pcc.sdf_idx_cnt = sdf_cnt;
-		} else {
-			tmp_pcc.adc_idx = atoi(entry);
-		}
-
-		ret = install_pcc_rules(tmp_pcc);
-        assert(ret >=0 );
-
-	}
-	num_pcc_filter = FIRST_FILTER_ID; /*Reset num_pcc_filter*/
-}
-
-void
-init_packet_filters(void)
-{
-	/* init pcc rule tables on dp*/
-	init_pcc_rules();
-
-	/* init dpn meter profile table before configuring pcc/adc rules*/
-	init_mtr_profile();
-
-	/* init dpn sdf rules table configuring on dp*/
-	init_sdf_rules();
-}
-
-/**
- * @brief  : Prints adc rule info
- * @param  : adc_rule, rule info to be printed
- * @return : Returns nothing
- */
-static void print_adc_rule(struct adc_rules adc_rule)
-{
-	LOG_MSG(LOG_DEBUG,"%-8u ", adc_rule.rule_id);
-	switch (adc_rule.sel_type) {
-	case DOMAIN_IP_ADDR:
-		LOG_MSG(LOG_DEBUG,"%-10s " IPV4_ADDR, "IP",
-			IPV4_ADDR_HOST_FORMAT(adc_rule.u.domain_ip.u.ipv4_addr));
-		break;
-	case DOMAIN_IP_ADDR_PREFIX:
-		LOG_MSG(LOG_DEBUG,"%-10s " IPV4_ADDR"/%d ", "IP_PREFIX",
-			IPV4_ADDR_HOST_FORMAT(adc_rule.u.domain_prefix.ip_addr.u.ipv4_addr),
-			adc_rule.u.domain_prefix.prefix);
-		break;
-	case DOMAIN_NAME:
-		LOG_MSG(LOG_DEBUG,"%-10s %-35s ", "DOMAIN", adc_rule.u.domain_name);
-		break;
-	default:
-		LOG_MSG(LOG_ERROR,"ERROR IN ADC RULE");
-	}
-}
-
-void
-parse_adc_rules(void)
-{
-	unsigned num_adc_rules = 0;
-	unsigned i = 0;
-	uint32_t rule_id = 1;
-	const char *entry = NULL;
-	struct rte_cfgfile *file = rte_cfgfile_load(ADC_RULE_FILE, 0);
-
-    assert(file != NULL);
-
-	entry = rte_cfgfile_get_entry(file, "GLOBAL", "NUM_ADC_RULES");
-
-    assert(entry != NULL);
-
-	num_adc_rules = atoi(entry);
-
-	for (i = 1; i <= num_adc_rules; ++i) {
-		char sectionname[64] = {0};
-		struct adc_rules tmp_adc = { 0 };
-		struct in_addr addr;
-
-		snprintf(sectionname, sizeof(sectionname),
-				"ADC_RULE_%u", i);
-
-		entry = rte_cfgfile_get_entry(file, sectionname,
-				"ADC_TYPE");
-
-        assert(entry != NULL);
-
-		tmp_adc.sel_type = atoi(entry);
-
-		switch (tmp_adc.sel_type) {
-			case DOMAIN_NAME:
-				entry = rte_cfgfile_get_entry(file, sectionname,
-						"DOMAIN");
-				if(entry)
-					strncpy(tmp_adc.u.domain_name, entry,
-							sizeof(tmp_adc.u.domain_name));
-				break;
-
-			case DOMAIN_IP_ADDR:
-				entry = rte_cfgfile_get_entry(file, sectionname,
-						"IP");
-
-				if (entry) {
-					inet_aton(entry, &addr);
-					tmp_adc.u.domain_ip.u.ipv4_addr = ntohl(addr.s_addr);
-					tmp_adc.u.domain_ip.iptype = IPTYPE_IPV4;
-				}
-				break;
-
-			case DOMAIN_IP_ADDR_PREFIX:
-				entry = rte_cfgfile_get_entry(file, sectionname,
-						"IP");
-
-				if (entry) {
-					inet_aton(entry, &addr);
-					tmp_adc.u.domain_ip.u.ipv4_addr = ntohl(addr.s_addr);
-					tmp_adc.u.domain_ip.iptype = IPTYPE_IPV4;
-				}
-
-				entry = rte_cfgfile_get_entry(file, sectionname,
-						"PREFIX");
-
-				if (entry)
-					tmp_adc.u.domain_prefix.prefix = atoi(entry);
-
-				break;
-
-			default:
-                assert(0);
-		}
-
-
-		/* Add Default rule */
-		adc_rule_id[rule_id - 1] = rule_id;
-		tmp_adc.rule_id = rule_id++;
-#ifdef OBSOLETE_APIS
-	struct dp_id dp_id = { .id = DPN_ID };
-		if (adc_entry_add(dp_id, tmp_adc) < 0)
-                assert(0);
-#endif
-		print_adc_rule(tmp_adc);
-
-	}
-	num_adc_rules = rule_id - 1;
-
+       memcpy(pf, &catch_all, sizeof(pkt_fltr));
 }
